% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{algo,BinSeg-method}
\alias{algo,BinSeg-method}
\alias{algo}
\title{Obtain the algorithm used in the analysis.}
\usage{
\S4method{algo}{BinSeg}(object)
}
\arguments{
\item{object}{A valid BinSeg object.}
}
\value{
A string with the algorithm used.
}
\description{
Obtain the algorithm used in the analysis.
}
