% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\docType{class}
\name{BinSeg-class}
\alias{BinSeg-class}
\title{An S4 class that represent a set of changepoint models.}
\description{
An S4 class that represent a set of changepoint models.
}
\section{Slots}{

\describe{
\item{\code{data}}{A numeric vector with the data used to perform the changepoint analysis.}

\item{\code{models_summary}}{A data.table with the summary of the segmentation models. It is only intended for internal usage.
It is recommended to use the other functions to interact with the changepoint model and obtain parameter estimations.}

\item{\code{algorithm}}{The algorithm string selected for the changepoint analysis.}

\item{\code{distribution}}{The distribution string selected for the changepoint analysis.}

\item{\code{min_seg_len}}{The value selected as the minimum segment length (default is 1).}

\item{\code{param_names}}{The parameters that will be estimated by the changepoint model according to the selected distribution.}
}}

\section{Warning}{

You should not create BinSeg objects by hand. You must provide the data to the BinSegModel function, which
will compute the changepoint model by using the Rcpp interface and then fix the output to create a BinSeg object and
return it.
}

\seealso{
BinSegModel for computing the changepoint model and obtaining a BinSeg object.
}
