\name{BinSeg-package}
\alias{BinSeg-package}
\alias{BinSeg}
\docType{package}
\title{\packageTitle{BinSeg}}
\description{\packageDescription{BinSeg}}
\section{Package Content}{\packageIndices{BinSeg}}
\author{\packageAuthor{BinSeg}}
\section{Maintainer}{\packageMaintainer{BinSeg}}
\keyword{package}
