% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInterface.R
\name{BinSegInfo}
\alias{BinSegInfo}
\title{Check available algorithms and distributions}
\usage{
BinSegInfo()
}
\value{
A List with two slots, one for Algorithms and one for Distributions. Each one of them is a Character Matrix
with a column for the parameter string and another one for the description.
}
\description{
This function allows to check which algorithms and distributions are implemented in the package. It provides the string
that must be passed as a parameter to the BinSeg method, along with a small description.
}
\section{Details}{

This method queries directly the C++ code in order to find which distributions are currently available, as well as the
name that must be passed to use them. The benefit is that no R code must be modified when a new Distribution or
Algorithm is added. Only by adding on C++ it is possible to register it on the return matrix of this method.
}

\seealso{
BinSegModel to perform changepoint analysis.
}
