% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{coef,BinSeg-method}
\alias{coef,BinSeg-method}
\title{Obtain segment data for a particular number of changepoints.}
\usage{
\S4method{coef}{BinSeg}(object, segments = seq_len(nrow(object@models_summary)))
}
\arguments{
\item{object}{A valid BinSeg object}

\item{segments}{The number of segments to use. Must be a numeric value. If you provide only one segment, it must be
accompanied by L (eg. 1L). Otherwise, if you want information about several models (ie. several number of changepoints
/ segments), provide a vector.}
}
\value{
A data.table with the following fixed columns:
#' \describe{
     \item{segments}{This identifies a specific model by the number of segments. For the model with five segments,
each of the five rows will have a value of 5 in this column.}
     \item{start}{The starting index of a segment relative to the input data vector number of data points. It is
unique for a partiuclar model (i.e. no two segments can start on the same index}
     \item{end}{The ending index of a segment}
}
Additionally, columns for each estimated parameter are added. However, this depends on the selected distribution. For
example, if the distribution is \code{meanvar_norm}, then two columns will be added, one for segment mean and another
one for segment variance.
}
\description{
This function allows to obtain segment start, end, and parameter estimation (mean, etc. depending on the
selected distribution) for the selected number of segments. Can compute segment data for several models at the same time.
Note that the models_summary produced by rcpp_binseg is used here in order to comopute several models without doing the
changepoint analysis again.
}
\seealso{
rcpp_binseg for the function that creates the models_summary matrix used in this function.
}
