% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{cpts,BinSeg-method}
\alias{cpts,BinSeg-method}
\title{Obtain the changepoints estimated by the analysis.}
\usage{
\S4method{cpts}{BinSeg}(object, ncpts = seq_len(nrow(object@models_summary)))
}
\arguments{
\item{object}{A valid BinSeg object.}

\item{nctps}{The number of changepoints to return (from 1 up to the number of changepoints indicated in BinSegModel).}
}
\value{
A numeric vector
}
\description{
Obtain the changepoints estimated by the analysis.
}
\section{Details}{

The changepoints are not sorted. Instead, they are arranged according to the decrease in cost. The first changepoint
will always be the last data point. From there, the next changepoint will be the one that decreases the most the overall
cost of the model, if there was only one changepoint, and so on.
}

