% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{dist,BinSeg-method}
\alias{dist,BinSeg-method}
\title{Obtain the distribution used in the analysis.}
\usage{
\S4method{dist}{BinSeg}(object)
}
\arguments{
\item{object}{A valid BinSeg object}
}
\value{
A string with the distribution used.
}
\description{
Obtain the distribution used in the analysis.
}
