% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{logLik,BinSeg-method}
\alias{logLik,BinSeg-method}
\title{Obtain the overall costs of the computed models, from 1 up to the selected number of changepoints.}
\usage{
\S4method{logLik}{BinSeg}(object, ncpts = seq_len(nrow(object@models_summary)))
}
\arguments{
\item{object}{A valid BinSeg object.}

\item{ncpts}{The number of changepoints to return (from 1 up to the number of changepoints indicated in BinSegModel).}
}
\value{
A numeric matrix with the overall cost at each number of changepoints
}
\description{
This function returns the overall cost for each one of the models. When you perform a BinSegModel, the models up to the
sepcified number of changepoints are computed. This functions returns the cost of each one of them. The first value is
the cost of the model without segmentation (remember the first changepoint is the last datapoint). Then, it tells the
overall cost with each addition of a segment.
}
