% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{plotDiagnostic,BinSeg-method}
\alias{plotDiagnostic,BinSeg-method}
\title{Plot a diagnostic of the models}
\usage{
\S4method{plotDiagnostic}{BinSeg}(object, ncpts = seq_len(nrow(object@models_summary)))
}
\arguments{
\item{object}{A valid BinSeg object}

\item{ncpts}{The number of changepoints to consider (Default is whole model)}
}
\value{
A ggplot with the graph of the costs.
}
\description{
This method creates a graph of the overall cost at each changepoint model up to the selected number of changepoints.
It allows to determine how many changepoints to consider and to visualize the cost decrease that each new segment brings.
}
